+++ ziti/util/identities.go
@@ -34,6 +34,7 @@ const (
 type RestClientConfig struct {
 	EdgeIdentities map[string]*RestClientEdgeIdentity `json:"edgeIdentities"`
 	Default        string                             `json:"default"`
+	Layout         int                                `json:"layout"`
 }
 
 func (self *RestClientConfig) GetIdentity() string {
@@ -184,6 +185,7 @@ func LoadRestClientConfig() (*RestClientConfig, string, error) {
 	if err != nil {
 		return nil, "", errors.Wrap(err, "couldn't get config dir while loading cli configuration")
 	}
+
 	configFile := filepath.Join(cfgDir, "ziti-cli.json")
 	_, err = os.Stat(configFile)
 	if err != nil {
@@ -193,6 +195,7 @@ func LoadRestClientConfig() (*RestClientConfig, string, error) {
 		}
 		return nil, "", errors.Wrapf(err, "error while statting config file %v", configFile)
 	}
+
 	result, err := os.ReadFile(configFile)
 	if err != nil {
 		return nil, "", errors.Wrapf(err, "error while reading config file %v", configFile)
@@ -218,7 +221,8 @@ func PersistRestClientConfig(config *RestClientConfig) error {
 	if err != nil {
 		return errors.Wrap(err, "couldn't get config dir while persisting cli configuration")
 	}
-	if err := os.MkdirAll(cfgDir, 0700); err != nil {
+
+	if err = os.MkdirAll(cfgDir, 0700); err != nil {
 		return errors.Wrapf(err, "unable to create config dir %v", cfgDir)
 	}